// project_member/src/types/types.ts
export type Aluno = {
  id: string;
  nome: string;
  cpf: string;
  email: string;
  telefone: string;
  cursoSelecionado: string;
  nivelEnsino: string;
  sexo: string;
  edital: string;
  projetos: string;
  vinculo: string;
  valorBolsa: string;
  valorOutro: string;
  banco: string;
  agencia: string;
  conta: string;
  dataCadastro: Date;
};